refactor: common routes to MainRoutes class
Separates main common routes to MainRoutes.
SplunkIntegration would only define its handler.

EVNT-601
